requirements.txt
Django
djangorestframework
drf-spectacular
ipython
requests
